reorganize files and folders for User and Auth
