+++ src/main/kotlin/io/github/soat7/myburguercontrol/domain/mapper/UserMapper.kt
@@ -2,8 +2,8 @@ package io.github.soat7.myburguercontrol.domain.mapper
 
 import io.github.soat7.myburguercontrol.domain.model.User
 import io.github.soat7.myburguercontrol.infrastructure.persistence.user.entity.UserEntity
-import io.github.soat7.myburguercontrol.infrastructure.rest.api.UserCreationRequest
-import io.github.soat7.myburguercontrol.infrastructure.rest.api.UserResponse
+import io.github.soat7.myburguercontrol.infrastructure.rest.user.api.UserCreationRequest
+import io.github.soat7.myburguercontrol.infrastructure.rest.user.api.UserResponse
 import java.util.UUID
 
 fun UserCreationRequest.toDomain() = User(
