+++ src/main/kotlin/io/github/soat7/myburguercontrol/domain/service/AuthenticationService.kt
@@ -4,8 +4,8 @@ import io.github.soat7.myburguercontrol.application.ports.inbound.Authentication
 import io.github.soat7.myburguercontrol.application.ports.inbound.CustomUserDetailsServicePort
 import io.github.soat7.myburguercontrol.application.ports.inbound.TokenServicePort
 import io.github.soat7.myburguercontrol.infrastructure.config.JwtProperties
-import io.github.soat7.myburguercontrol.infrastructure.rest.api.AuthRequest
-import io.github.soat7.myburguercontrol.infrastructure.rest.api.AuthResponse
+import io.github.soat7.myburguercontrol.infrastructure.rest.user.api.AuthRequest
+import io.github.soat7.myburguercontrol.infrastructure.rest.user.api.AuthResponse
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import java.util.Date
