+++ src/main/kotlin/io/github/soat7/myburguercontrol/infrastructure/config/BeanConfiguration.kt
@@ -5,8 +5,6 @@ import io.github.soat7.myburguercontrol.application.ports.inbound.CustomUserDeta
 import io.github.soat7.myburguercontrol.application.ports.inbound.TokenServicePort
 import io.github.soat7.myburguercontrol.application.ports.outbound.CustomerDatabasePort
 import io.github.soat7.myburguercontrol.application.ports.outbound.ProductDatabasePort
-//import io.github.soat7.myburguercontrol.application.ports.outbound.PaymentIntegrationPort
-//import io.github.soat7.myburguercontrol.application.ports.outbound.ItemDatabasePort
 import io.github.soat7.myburguercontrol.application.ports.outbound.UserDatabasePort
 import io.github.soat7.myburguercontrol.domain.service.AuthenticationService
 import io.github.soat7.myburguercontrol.domain.service.CustomUserDetailsService
@@ -15,8 +13,6 @@ import io.github.soat7.myburguercontrol.domain.service.ProductService
 import io.github.soat7.myburguercontrol.domain.service.TokenService
 import io.github.soat7.myburguercontrol.domain.service.UserService
 import io.github.soat7.myburguercontrol.infrastructure.persistence.user.repository.UserRepository
-//import io.github.soat7.myburguercontrol.domain.service.PaymentService
-//import io.github.soat7.myburguercontrol.domain.service.ItemService
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.authentication.AuthenticationManager
