--- src/main/kotlin/io/github/soat7/myburguercontrol/infrastructure/rest/api/AuthResponse.kt
@@ -1,5 +1,0 @@
-package io.github.soat7.myburguercontrol.infrastructure.rest.api
-
-data class AuthResponse(
-    val accessToken: String
-)
