+++ src/main/kotlin/io/github/soat7/myburguercontrol/infrastructure/rest/user/api/AuthRequest.kt
@@ -1,4 +1,4 @@
-package io.github.soat7.myburguercontrol.infrastructure.rest.api
+package io.github.soat7.myburguercontrol.infrastructure.rest.user.api
 
 data class AuthRequest(
     val cpf: String,
