+++ src/main/kotlin/io/github/soat7/myburguercontrol/infrastructure/rest/user/api/AuthResponse.kt
@@ -1,0 +1,5 @@
+package io.github.soat7.myburguercontrol.infrastructure.rest.user.api
+
+data class AuthResponse(
+    val accessToken: String
+)
