+++ src/main/kotlin/io/github/soat7/myburguercontrol/infrastructure/rest/user/api/UserResponse.kt
@@ -1,4 +1,4 @@
-package io.github.soat7.myburguercontrol.infrastructure.rest.api
+package io.github.soat7.myburguercontrol.infrastructure.rest.user.api
 
 import io.github.soat7.myburguercontrol.domain.model.Role
 import java.util.UUID
