+++ src/test/kotlin/io/github/soat7/myburguercontrol/fixtures/AuthFixtures.kt
@@ -1,6 +1,6 @@
 package io.github.soat7.myburguercontrol.fixtures
 
-import io.github.soat7.myburguercontrol.infrastructure.rest.api.AuthRequest
+import io.github.soat7.myburguercontrol.infrastructure.rest.user.api.AuthRequest
 
 object AuthFixtures {
