+++ src/test/kotlin/io/github/soat7/myburguercontrol/infrastructure/rest/AuthIT.kt
@@ -1,4 +1,4 @@
-package io.github.soat7.myburguercontrol.application.rest
+package io.github.soat7.myburguercontrol.infrastructure.rest
 
 import io.github.soat7.myburguercontrol.base.BaseIntegrationTest
 import io.github.soat7.myburguercontrol.fixtures.AuthFixtures
@@ -10,9 +10,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.function.Executable
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.web.client.postForEntity
-import org.springframework.http.HttpStatus
 import org.springframework.security.crypto.password.PasswordEncoder
-import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
 class AuthIT : BaseIntegrationTest() {
